Magnetometer (LSM9DS1) Incorporation
Adafruit recently updated their libraries for the LSM9DS1 Magnetometer chip we are using, so it now functions properly. This code now additionally finds the location of a magnet to the magnetometer, which we should use to get a weighted average of the location of the pen, hopefully improving accuracy.